r fjerner amplitude fra eventlogging
