+++ package.json
@@ -14,7 +14,6 @@
     },
     "dependencies": {
         "@alheimsins/virkedager": "2.2.0",
-        "@amplitude/analytics-browser": "2.23.1",
         "@grafana/faro-web-sdk": "1.19.0",
         "@grafana/faro-web-tracing": "1.19.0",
         "@navikt/aksel-icons": "7.28.1",
