--- src/components/init-amplitude.tsx
@@ -1,18 +1,0 @@
-'use client';
-
-import { useEffect } from 'react';
-import { initAmplitude } from '@/lib/amplitude';
-
-interface Props {
-    apiKey: string;
-}
-
-const InitAmplitude = ({ apiKey }: Props) => {
-    useEffect(() => {
-        initAmplitude({ apiKey });
-    }, [apiKey]);
-
-    return null;
-};
-
-export default InitAmplitude;
