+++ src/components/init-tracker.tsx
@@ -1,0 +1,18 @@
+'use client';
+
+import { useEffect } from 'react';
+import { initTracker } from '@/lib/tracker';
+
+interface Props {
+    apiKey: string;
+}
+
+const InitTracker = ({ apiKey }: Props) => {
+    useEffect(() => {
+        initTracker({ apiKey });
+    }, [apiKey]);
+
+    return null;
+};
+
+export default InitTracker;
