+++ src/components/registrerings-oversikt.tsx
@@ -4,7 +4,7 @@ import useSprak from '../hooks/useSprak';
 import { FormProgress, Heading } from '@navikt/ds-react';
 import { SkjemaSide, SkjemaState } from '@/model/skjema';
 import { lagHentTekstForSprak } from '@navikt/arbeidssokerregisteret-utils';
-import { loggAktivitet } from '@/lib/amplitude';
+import { loggAktivitet } from '@/lib/tracker';
 
 const TEKSTER = {
     nb: {
