+++ src/components/skjema/fullfor-registrering-knapp-ny-inngang.tsx
@@ -14,7 +14,7 @@ import { fetcher as api } from '../../lib/api-utils';
 import hentKvitteringsUrl from '../../lib/hent-kvitterings-url';
 import { FeilmeldingGenerell } from '../feilmeldinger/feilmeldinger';
 import { SkjemaState } from '@/model/skjema';
-import { loggAktivitet, loggFlyt } from '@/lib/amplitude';
+import { loggAktivitet, loggFlyt } from '@/lib/tracker';
 import useSprak from '@/hooks/useSprak';
 
 interface FullforKnappProps {
